typography: Remove all explicit font feature settings.
Originally we used to hardcode Manrope as *the* Helium typeface, which
makes sense for us to enforce these font features since only tauOS used
to be the only user of Helium.

Now that we're using Helium in pretty much our whole desktop apps set
including Ultramarine (which is less opinionated), this font feature
toggle actually makes no sense at all, and now that we're switching to
either Inter or Geist as the new typeface, this feature makes no sense
and also breaks Inter.

This commit removes the font-feature settings to restore sensible font
defaults for other typefaces